requirements.txt
Django==2.0.1
django-admin-bootstrapped==2.5.7
facebook-sdk==2.0.0
Pillow==5.0.0
